refactor: extract core helpers into internal package
